add tests and minor refactoring
