+++ jhub_apps/tests/test_api.py
@@ -5,6 +5,7 @@ from unittest.mock import patch
 import pytest
 
 from jhub_apps.hub_client.hub_client import HubClient
+from jhub_apps.service.utils import get_shared_servers
 from jhub_apps.spawner.types import FRAMEWORKS
 from jhub_apps.tests.constants import MOCK_USER
 
@@ -142,6 +143,43 @@ def test_api_update_server(edit_server, client):
     )
     assert response.status_code == 200
     assert response.json() == create_server_response
+
+
+@patch.object(HubClient, "get_users")
+@patch.object(HubClient, "get_shared_servers")
+def test_shared_server_filtering(hub_get_shared_servers, get_users):
+    current_hub_user = {"name": "fakeuser"}
+    get_users.return_value = [
+        {
+            "servers": {
+                '': {'name': ''},
+                "panel-12": {"name": "panel-12"}
+            }
+        },
+        {
+            "servers": {
+                '': {'name': ''},
+                "panel-34": {"name": "panel-34", "fullname": "panel shared 34"}
+            }
+        },
+        {
+            "servers": {
+                '': {'name': ''},
+                "panel-56": {"name": "panel-56", "fullname": "panel shared server"}
+            }
+        }
+    ]
+    hub_get_shared_servers.return_value = [
+        {"server": {"name": "panel-56"}},
+        {"server": {"name": "panel-34"}},
+    ]
+    shared_servers = get_shared_servers(current_hub_user)
+    assert shared_servers == [
+        {"name": "panel-34", "fullname": "panel shared 34"},
+        {"name": "panel-56", "fullname": "panel shared server"}
+    ]
+    hub_get_shared_servers.assert_called_once_with()
+    get_users.assert_called_once_with()
 
 
 def test_api_frameworks(client):
